+++ argopy/__init__.py
@@ -26,11 +26,12 @@ from . import tutorial  # noqa: E402
 from . import utilities  # noqa: E402
 from . import stores  # noqa: E402
 from . import errors  # noqa: E402
-from . import plotters  # noqa: E402
+from . import plot  # noqa: E402
+from .plot import dashboard  # noqa: E402
 from .utilities import show_versions, show_options, clear_cache, lscache, TopoFetcher  # noqa: E402
 from .utilities import monitor_status as status  # noqa: E402
 from .options import set_options  # noqa: E402
-from .plotters import open_dashboard as dashboard  # noqa: E402
+
 
 #
 __all__ = (
@@ -50,7 +51,8 @@ __all__ = (
     # Sub-packages,
     "utilities",
     "errors",
-    "plotters",
+    "plot",
+    # "plotters",  # Deprec, to be removed after 0.1.13
     "stores",
     "tutorial",
     # Constants