+++ argopy/fetchers.py
@@ -17,13 +17,16 @@ import logging
 
 from argopy.options import OPTIONS, _VALIDATORS
 from .errors import InvalidFetcherAccessPoint, InvalidFetcher
+
 from .utilities import (
     list_available_data_src, list_available_index_src,
     is_box, is_indexbox,
-    check_wmo, check_cyc
+    check_wmo, check_cyc,
+    get_coriolis_profile_id
 )
-from .plotters import plot_trajectory, bar_plot, open_sat_altim_report
 from argopy.stores import filestore
+from .plot import plot_trajectory, bar_plot, open_sat_altim_report
+
 
 AVAILABLE_DATA_SOURCES = list_available_data_src()
 AVAILABLE_INDEX_SOURCES = list_available_index_src()
@@ -283,7 +286,12 @@ class ArgoDataFetcher:
                 np.min(this_ds['TIME'].values), np.max(this_ds['TIME'].values)]
 
     def dashboard(self, **kw):
-        """ Open access point dashboard """
+        """Open access point dashboard.
+
+            See Also
+            --------
+            :class:`argopy.dashboard`
+        """
         try:
             return self.fetcher.dashboard(**kw)
         except Exception:
@@ -445,7 +453,7 @@ class ArgoDataFetcher:
             )
         return self.load().data.to_dataframe(**kwargs)
 
-    def to_index(self, full: bool = False):
+    def to_index(self, full: bool = False, coriolis_id = False):
         """ Create an index of Argo data, fetch data if necessary
 
             Build an Argo-like index of profiles from fetched data.
@@ -477,11 +485,13 @@ class ArgoDataFetcher:
                     ds.drop_vars(set(ds.data_vars) - set(["PLATFORM_NUMBER"]))
                     .to_dataframe()
                 )
+
             df = (
                 df.reset_index()
                 .rename(
                     columns={
                         "PLATFORM_NUMBER": "wmo",
+                        "CYCLE_NUMBER": "cyc",
                         "LONGITUDE": "longitude",
                         "LATITUDE": "latitude",
                         "TIME": "date",
@@ -489,8 +499,13 @@ class ArgoDataFetcher:
                 )
                 .drop(columns="N_PROF")
             )
-            df = df[["date", "latitude", "longitude", "wmo"]]
-
+            df = df[["date", "latitude", "longitude", "wmo", "cyc"]]
+            if coriolis_id:
+                df['id'] = None
+                def fc(row):
+                    row['id'] = get_coriolis_profile_id(row['wmo'], row['cyc'])['ID'].values[0]
+                    return row
+                df = df.apply(fc, axis=1)
         else:
             # Instantiate and load an IndexFetcher:
             index_loader = ArgoIndexFetcher(mode=self._mode,
@@ -546,7 +561,7 @@ class ArgoDataFetcher:
         if not self._loaded or force:
             # Fetch measurements:
             self._data = self.to_xarray(**kwargs)
-            # Next 2 lines must come before ._index because to_index() calls back on .load() to read .data
+            # Next 2 lines must come before ._index because to_index(full=False) calls back on .load() to read .data
             self._request = self.__repr__()  # Save definition of loaded data
             self._loaded = True
             # Extract measurements index from data:
@@ -567,7 +582,8 @@ class ArgoDataFetcher:
 
             Parameters
             ----------
-            ptype: {'trajectory',' profiler', 'dac', 'qc_altimetry}, default: 'trajectory'
+            ptype: str, optional, default: 'trajectory'
+                Plot type, one of the following: 'trajectory',' profiler', 'dac', 'qc_altimetry'.
 
             Returns
             -------