+++ argopy/tests/test_plotters.py
@@ -4,6 +4,7 @@ We test plotting functions from IndexFetcher and DataFetcher
 """
 import pytest
 import importlib
+import logging
 
 import argopy
 from argopy.errors import InvalidDashboard
@@ -16,8 +17,9 @@ from utils import (
     has_matplotlib,
     has_seaborn,
     has_cartopy,
+    has_ipython,
 )
-from argopy.plotters import bar_plot, plot_trajectory, open_sat_altim_report
+from ..plot import bar_plot, plot_trajectory, open_sat_altim_report
 from argopy import IndexFetcher as ArgoIndexFetcher
 from argopy import DataFetcher as ArgoDataFetcher
 
@@ -27,36 +29,62 @@ if has_matplotlib:
 if has_cartopy:
     import cartopy
 
+log = logging.getLogger("argopy.tests.plot")
 
-@requires_connection
-def test_invalid_dashboard():
+
+@pytest.mark.parametrize("board_type", ["invalid", "argovis", "op", "ocean-ops", "coriolis"], indirect=False)
+def test_invalid_dashboard(board_type):
+    # Test types without 'base'
     with pytest.raises(InvalidDashboard):
-        argopy.dashboard(wmo=5904797, type="invalid_service")
+        argopy.dashboard(type=board_type, url_only=True)
+
+@pytest.mark.parametrize("board_type", ["op", "ocean-ops", "coriolis"], indirect=False)
+def test_invalid_dashboard_profile(board_type):
+    # Test types without 'cyc'
+    with pytest.raises(InvalidDashboard):
+        argopy.dashboard(6902755, 12, type=board_type, url_only=True)
+
+@pytest.mark.parametrize("board_type", ["data", "meta", "ea", "eric", "bgc"], indirect=False)
+def test_valid_dashboard(board_type):
+    # Test types with 'base'
+    assert isinstance(argopy.dashboard(type=board_type, url_only=True), str)
+
+@pytest.mark.parametrize("board_type", ["data", "meta", "ea", "eric", "argovis", "op", "ocean-ops", "bgc"], indirect=False)
+def test_valid_dashboard_float(board_type):
+    # Test types with 'wmo' (should be all)
+    assert isinstance(argopy.dashboard(5904797, type=board_type, url_only=True), str)
+
+@pytest.mark.parametrize("board_type", ["data", "meta", "ea", "eric", "argovis", "bgc"], indirect=False)
+def test_valid_dashboard_profile(board_type):
+    # Test types with 'cyc'
+    assert isinstance(argopy.dashboard(5904797, 12, type=board_type, url_only=True), str)
 
 
 @requires_ipython
 @requires_connection
-def test_valid_dashboard():
+def test_valid_dashboard_ipython_output():
     import IPython
+
+    dsh = argopy.dashboard()
+    assert isinstance(dsh, IPython.lib.display.IFrame)
+
     dsh = argopy.dashboard(wmo=5904797)
     assert isinstance(dsh, IPython.lib.display.IFrame)
 
+    dsh = argopy.dashboard(wmo=5904797, cyc=3)
+    assert isinstance(dsh, IPython.lib.display.IFrame)
 
-@requires_ipython
+
 @requires_connection
 def test_open_sat_altim_report():
-    import IPython
-    dsh = open_sat_altim_report(WMO=5904797, embed='slide')
-    assert isinstance(dsh(0), IPython.display.Image)
+    if has_ipython:
+        import IPython
 
     dsh = open_sat_altim_report(WMO=5904797, embed='dropdown')
-    assert isinstance(dsh(5904797), IPython.display.Image)
-
-    assert argopy.plotters.open_sat_altim_report(WMO=5904797, embed='list') is None
-
-    dsh = open_sat_altim_report(WMO=5904797, embed=None)
-    assert isinstance(dsh, dict)
-    assert 5904797 in dsh
+    if has_ipython:
+        assert isinstance(dsh(5904797), IPython.display.Image)
+    else:
+        assert isinstance(dsh, dict)
 
 
 @requires_gdac