+++ argopy/utilities.py
@@ -48,7 +48,7 @@ from .errors import (
     InvalidFetcherAccessPoint,
     InvalidOption,
     InvalidDatasetStructure,
-    FileSystemHasNoCache
+    FileSystemHasNoCache,
 )
 
 try:
@@ -397,7 +397,7 @@ def list_multiprofile_file_variables():
 
 
 def check_localftp(path, errors: str = "ignore"):
-    """ Check if the path has the expected GDAC ftp structure
+    """Check if the path has the expected GDAC ftp structure.
 
         Check if the path is structured like:
         .
@@ -413,8 +413,8 @@ def check_localftp(path, errors: str = "ignore"):
         ----------
         path: str
             Path name to check
-        errors: str
-            "ignore" or "raise" (or "warn"
+        errors: str, optional
+            "ignore" or "raise" (or "warn")
 
         Returns
         -------
@@ -708,9 +708,7 @@ def isAPIconnected(src="erddap", data=True):
     else:
         list_src = list_available_index_src()
 
-    if src in list_src and getattr(
-        list_src[src], "api_server_check", None
-    ):
+    if src in list_src and getattr(list_src[src], "api_server_check", None):
         if "localftp" in src:
             # This is a special case because the source here is a local folder
             result = check_localftp(OPTIONS["local_ftp"])
@@ -721,7 +719,9 @@ def isAPIconnected(src="erddap", data=True):
         raise InvalidFetcher
 
 
-def erddap_ds_exists(ds: str = "ArgoFloats", erddap: str = 'https://www.ifremer.fr/erddap') -> bool:
+def erddap_ds_exists(
+    ds: str = "ArgoFloats", erddap: str = "https://www.ifremer.fr/erddap"
+) -> bool:
     """ Check if a dataset exists on a remote erddap server
     return a bool
 
@@ -796,7 +796,7 @@ def fetch_status(stdout: str = "html", insert: bool = True):
         if getattr(mod, "api_server_check", None):
             # status = isconnected(mod.api_server_check)
             status = isAPIconnected(api)
-            if api == 'localftp' and OPTIONS['local_ftp'] == '-':
+            if api == "localftp" and OPTIONS["local_ftp"] == "-":
                 message = "ok" if status else "path undefined !"
             else:
                 # message = "up" if status else "down"
@@ -957,7 +957,9 @@ def linear_interpolation_remap(
             vectorize=True,
             dask="parallelized",
             output_dtypes=[data.dtype],
-            dask_gufunc_kwargs={'output_sizes': {output_dim: len(z_regridded[z_regridded_dim])}},
+            dask_gufunc_kwargs={
+                "output_sizes": {output_dim: len(z_regridded[z_regridded_dim])}
+            },
         )
     else:
         kwargs = dict(
@@ -1051,7 +1053,7 @@ class Chunker:
     def _split(self, lst, n=1):
         """Yield successive n-sized chunks from lst"""
         for i in range(0, len(lst), n):
-            yield lst[i: i + n]
+            yield lst[i : i + n]
 
     def _split_list_bychunknb(self, lst, n=1):
         """Split list in n-imposed chunks of similar size
@@ -1062,7 +1064,7 @@ class Chunker:
         for i in self._split(lst, siz):
             res.append(i)
         if len(res) > n:
-            res[n - 1::] = [reduce(lambda i, j: i + j, res[n - 1::])]
+            res[n - 1 : :] = [reduce(lambda i, j: i + j, res[n - 1 : :])]
         return res
 
     def _split_list_bychunksize(self, lst, max_size=1):
@@ -1098,7 +1100,7 @@ class Chunker:
                     this_box[i_right] = right
                     boxes.append(this_box)
         elif "t" in d:
-            dates = pd.to_datetime(large_box[i_left: i_right + 1])
+            dates = pd.to_datetime(large_box[i_left : i_right + 1])
             date_bounds = [
                 d.strftime("%Y%m%d%H%M%S")
                 for d in pd.date_range(dates[0], dates[1], periods=n + 1)
@@ -1256,23 +1258,20 @@ def format_oneline(s, max_width=65):
         if q == 0:
             return "".join([s[0:n], padding, s[-n:]])
         else:
-            return "".join([s[0: n + 1], padding, s[-n:]])
+            return "".join([s[0 : n + 1], padding, s[-n:]])
     else:
         return s
 
 
 def is_indexbox(box: list, errors="raise"):
     """ Check if this array matches a 2d or 3d index box definition
-
         box = [lon_min, lon_max, lat_min, lat_max]
     or:
         box = [lon_min, lon_max, lat_min, lat_max, datim_min, datim_max]
-
     Parameters
     ----------
     box: list
     errors: 'raise'
-
     Returns
     -------
     bool
@@ -1360,6 +1359,7 @@ def is_box(box: list, errors="raise"):
     -------
     bool
     """
+
     def is_dateconvertible(d):
         try:
             pd.to_datetime(d)
@@ -1450,7 +1450,9 @@ def is_list_of_datasets(lst):
 
 def is_list_equal(lst1, lst2):
     """ Return true if 2 lists contain same elements"""
-    return len(lst1) == len(lst2) and len(lst1) == sum([1 for i, j in zip(lst1, lst2) if i == j])
+    return len(lst1) == len(lst2) and len(lst1) == sum(
+        [1 for i, j in zip(lst1, lst2) if i == j]
+    )
 
 
 def check_wmo(lst, errors="raise"):
@@ -1538,14 +1540,12 @@ def is_wmo(lst, errors="raise"):  # noqa: C901
 
 def check_cyc(lst, errors="raise"):
     """ Validate a CYC option and returned it as a list of integers
-
     Parameters
     ----------
     cyc: int
         CYC must be an integer or an iterable with elements that can be casted as positive integers
     errors: {'raise', 'warn', 'ignore'}
         Possibly raises a ValueError exception or UserWarning, otherwise fails silently.
-
     Returns
     -------
     list(int)
@@ -1565,14 +1565,12 @@ def check_cyc(lst, errors="raise"):
 
 def is_cyc(lst, errors="raise"):  # noqa: C901
     """ Check if a CYC is valid
-
     Parameters
     ----------
     cyc: int, list(int), array(int)
         CYC must be a single or a list of at most 4 digit positive numbers
     errors: {'raise', 'warn', 'ignore'}
         Possibly raises a ValueError exception or UserWarning, otherwise fails silently.
-
     Returns
     -------
     bool
@@ -1640,16 +1638,6 @@ def check_index_cols(column_names: list, convention: str = 'ar_index_global_prof
     else:
         return column_names
 
-# def docstring(value):
-#     """Replace one function docstring
-#
-#         To be used as a decorator
-#     """
-#     def _doc(func):
-#         func.__doc__ = value
-#         return func
-#     return _doc
-
 
 def warnUnless(ok, txt):
     """ Decorator to raise warning unless condition is True
@@ -1663,6 +1651,7 @@ def warnUnless(ok, txt):
     txt: str
         Text to display in the warning
     """
+
     def inner(fct):
         def wrapper(*args, **kwargs):
             warnings.warn("%s %s" % (fct.__name__, txt))
@@ -1708,7 +1697,9 @@ def modified_environ(*remove, **update):
         [env.pop(k) for k in remove_after]
 
 
-def toYearFraction(this_date: pd._libs.tslibs.timestamps.Timestamp = pd.to_datetime('now', utc=True)):
+def toYearFraction(
+    this_date: pd._libs.tslibs.timestamps.Timestamp = pd.to_datetime("now", utc=True)
+):
     """ Compute decimal year, robust to leap years, precision to the second
 
     Compute the fraction of the year a given timestamp corresponds to.
@@ -1733,7 +1724,9 @@ def toYearFraction(this_date: pd._libs.tslibs.timestamps.Timestamp = pd.to_datet
         startOfThisYear = pd.to_datetime("%i-01-01T00:00:00.000" % this_date.year, utc=True)
     else:
         startOfThisYear = pd.to_datetime("%i-01-01T00:00:00.000" % this_date.year)
-    yearDuration_sec = (startOfThisYear + pd.offsets.DateOffset(years=1) - startOfThisYear).total_seconds()
+    yearDuration_sec = (
+        startOfThisYear + pd.offsets.DateOffset(years=1) - startOfThisYear
+    ).total_seconds()
 
     yearElapsed_sec = (this_date - startOfThisYear).total_seconds()
     fraction = yearElapsed_sec / yearDuration_sec
@@ -1759,9 +1752,11 @@ def YearFraction_to_datetime(yf: float):
     fraction = np.round(fraction, 10)
 
     startOfThisYear = pd.to_datetime("%i-01-01T00:00:00" % year)
-    yearDuration_sec = (startOfThisYear + pd.offsets.DateOffset(years=1) - startOfThisYear).total_seconds()
-    yearElapsed_sec = pd.Timedelta(fraction * yearDuration_sec, unit='s')
-    return pd.to_datetime(startOfThisYear + yearElapsed_sec, unit='s')
+    yearDuration_sec = (
+        startOfThisYear + pd.offsets.DateOffset(years=1) - startOfThisYear
+    ).total_seconds()
+    yearElapsed_sec = pd.Timedelta(fraction * yearDuration_sec, unit="s")
+    return pd.to_datetime(startOfThisYear + yearElapsed_sec, unit="s")
 
 
 def wrap_longitude(grid_long):
@@ -1831,18 +1826,18 @@ def wmo2box(wmo_id: int):
     #
     dd = 10
     if quadrant in [1, 3]:
-        lon_min = nearest_to_the_Prime_Meridian*dd
-        lon_max = nearest_to_the_Prime_Meridian*dd+dd
+        lon_min = nearest_to_the_Prime_Meridian * dd
+        lon_max = nearest_to_the_Prime_Meridian * dd + dd
     elif quadrant in [5, 7]:
-        lon_min = -nearest_to_the_Prime_Meridian*dd-dd
-        lon_max = -nearest_to_the_Prime_Meridian*dd
+        lon_min = -nearest_to_the_Prime_Meridian * dd - dd
+        lon_max = -nearest_to_the_Prime_Meridian * dd
 
     if quadrant in [1, 7]:
-        lat_min = nearest_to_the_Equator_latitude*dd
-        lat_max = nearest_to_the_Equator_latitude*dd+dd
+        lat_min = nearest_to_the_Equator_latitude * dd
+        lat_max = nearest_to_the_Equator_latitude * dd + dd
     elif quadrant in [3, 5]:
-        lat_min = -nearest_to_the_Equator_latitude*dd-dd
-        lat_max = -nearest_to_the_Equator_latitude*dd
+        lat_min = -nearest_to_the_Equator_latitude * dd - dd
+        lat_max = -nearest_to_the_Equator_latitude * dd
 
     box = [lon_min, lon_max, lat_min, lat_max]
     return box
@@ -1863,7 +1858,9 @@ def groupby_remap(z, data, z_regridded,   # noqa C901
         x = x[~idx]
         y = y[~idx]
 
-        ifound = np.digitize(x, target_values, right=right)  # ``bins[i-1] <= x < bins[i]``
+        ifound = np.digitize(
+            x, target_values, right=right
+        )  # ``bins[i-1] <= x < bins[i]``
         ifound -= 1  # Because digitize returns a 1-based indexing, we need to remove 1
         y_binned = np.ones_like(target_values) * np.nan
 
@@ -1872,27 +1869,29 @@ def groupby_remap(z, data, z_regridded,   # noqa C901
             iselect = ix[-1]
 
             # Map to y value at specific x index in the bin:
-            if select == 'shallow':
+            if select == "shallow":
                 iselect = iselect[0]  # min/shallow
                 mapped_value = y[iselect]
-            elif select == 'deep':
+            elif select == "deep":
                 iselect = iselect[-1]  # max/deep
                 mapped_value = y[iselect]
-            elif select == 'middle':
-                iselect = iselect[np.where(x[iselect] >= np.median(x[iselect]))[0][0]]  # median/middle
+            elif select == "middle":
+                iselect = iselect[
+                    np.where(x[iselect] >= np.median(x[iselect]))[0][0]
+                ]  # median/middle
                 mapped_value = y[iselect]
-            elif select == 'random':
+            elif select == "random":
                 iselect = iselect[np.random.randint(len(iselect))]
                 mapped_value = y[iselect]
 
             # or Map to y statistics in the bin:
-            elif select == 'mean':
+            elif select == "mean":
                 mapped_value = np.nanmean(y[iselect])
-            elif select == 'min':
+            elif select == "min":
                 mapped_value = np.nanmin(y[iselect])
-            elif select == 'max':
+            elif select == "max":
                 mapped_value = np.nanmax(y[iselect])
-            elif select == 'median':
+            elif select == "median":
                 mapped_value = np.median(y[iselect])
 
             else:
@@ -1922,7 +1921,9 @@ def groupby_remap(z, data, z_regridded,   # noqa C901
             vectorize=True,
             dask="parallelized",
             output_dtypes=[data.dtype],
-            dask_gufunc_kwargs={'output_sizes': {output_dim: len(z_regridded[z_regridded_dim])}},
+            dask_gufunc_kwargs={
+                "output_sizes": {output_dim: len(z_regridded[z_regridded_dim])}
+            },
         )
     else:
         kwargs = dict(
@@ -1935,11 +1936,13 @@ def groupby_remap(z, data, z_regridded,   # noqa C901
         )
     remapped = xr.apply_ufunc(_subsample_bins, z, data, z_regridded, **kwargs)
 
-    remapped.coords[output_dim] = z_regridded.rename({z_regridded_dim: output_dim}).coords[output_dim]
+    remapped.coords[output_dim] = z_regridded.rename(
+        {z_regridded_dim: output_dim}
+    ).coords[output_dim]
     return remapped
 
 
-class TopoFetcher():
+class TopoFetcher:
     """ Fetch topographic data through an ERDDAP server for an ocean rectangle
 
     Example:
@@ -1950,23 +1953,23 @@ class TopoFetcher():
 
     """
 
-    class ERDDAP():
-        def __init__(self, server: str, protocol: str = 'tabledap'):
+    class ERDDAP:
+        def __init__(self, server: str, protocol: str = "tabledap"):
             self.server = server
             self.protocol = protocol
-            self.response = 'nc'
-            self.dataset_id = ''
-            self.constraints = ''
+            self.response = "nc"
+            self.dataset_id = ""
+            self.constraints = ""
 
     def __init__(
-            self,
-            box: list,
-            ds: str = "gebco",
-            cache: bool = False,
-            cachedir: str = "",
-            api_timeout: int = 0,
-            stride: list = [1, 1],
-            **kwargs,
+        self,
+        box: list,
+        ds: str = "gebco",
+        cache: bool = False,
+        cachedir: str = "",
+        api_timeout: int = 0,
+        stride: list = [1, 1],
+        **kwargs,
     ):
         """ Instantiate an ERDDAP topo data fetcher
 
@@ -1987,25 +1990,25 @@ class TopoFetcher():
         """
         from .stores import httpstore
         timeout = OPTIONS["api_timeout"] if api_timeout == 0 else api_timeout
-        self.fs = httpstore(cache=cache, cachedir=cachedir, timeout=timeout, size_policy='head')
+        self.fs = httpstore(
+            cache=cache, cachedir=cachedir, timeout=timeout, size_policy="head"
+        )
         self.definition = "Erddap topographic data fetcher"
 
         self.BOX = box
         self.stride = stride
         if ds == "gebco":
             self.definition = "NOAA erddap gebco data fetcher for a space region"
-            self.server = 'https://coastwatch.pfeg.noaa.gov/erddap'
-            self.server_name = 'NOAA'
-            self.dataset_id = 'gebco'
+            self.server = "https://coastwatch.pfeg.noaa.gov/erddap"
+            self.server_name = "NOAA"
+            self.dataset_id = "gebco"
 
         self._init_erddap()
 
     def _init_erddap(self):
         # Init erddap
         self.erddap = self.ERDDAP(server=self.server, protocol="griddap")
-        self.erddap.response = (
-            "nc"
-        )
+        self.erddap.response = "nc"
 
         if self.dataset_id == "gebco":
             self.erddap.dataset_id = "GEBCO_2020"
@@ -2052,17 +2055,19 @@ class TopoFetcher():
 
     def define_constraints(self):
         """ Define request constraints """
-        # Eg: https://coastwatch.pfeg.noaa.gov/erddap/griddap/GEBCO_2020.nc?elevation%5B(34):5:(42)%5D%5B(-21):7:(-12)%5D
-        self.erddap.constraints = "%s(%0.2f):%i:(%0.2f)%s%s(%0.2f):%i:(%0.2f)%s" % ("%5B",
-                                                                                    self.BOX[2],
-                                                                                    self.stride[1],
-                                                                                    self.BOX[3],
-                                                                                    "%5D",
-                                                                                    "%5B",
-                                                                                    self.BOX[0],
-                                                                                    self.stride[0],
-                                                                                    self.BOX[1],
-                                                                                    "%5D")
+        #        Eg: https://coastwatch.pfeg.noaa.gov/erddap/griddap/GEBCO_2020.nc?elevation%5B(34):5:(42)%5D%5B(-21):7:(-12)%5D
+        self.erddap.constraints = "%s(%0.2f):%i:(%0.2f)%s%s(%0.2f):%i:(%0.2f)%s" % (
+            "%5B",
+            self.BOX[2],
+            self.stride[1],
+            self.BOX[3],
+            "%5D",
+            "%5B",
+            self.BOX[0],
+            self.stride[0],
+            self.BOX[1],
+            "%5D",
+        )
         return None
 
     #     @property
@@ -2122,7 +2127,7 @@ class TopoFetcher():
         """
         return [self.get_url()]
 
-    def to_xarray(self, errors: str = 'ignore'):
+    def to_xarray(self, errors: str = "ignore"):
         """ Load Topographic data and return a xarray.DataSet """
 
         # Download data
@@ -2131,7 +2136,7 @@ class TopoFetcher():
 
         return ds
 
-    def load(self, errors: str = 'ignore'):
+    def load(self, errors: str = "ignore"):
         """ Load Topographic data and return a xarray.DataSet """
         return self.to_xarray(errors=errors)
 
@@ -2677,3 +2682,266 @@ class Registry(UserList):
     def copy(self):
         """Return a shallow copy of the registry"""
         return self.__copy__()
+
+
+def get_coriolis_profile_id(WMO, CYC=None):
+    """ Return a :class:`pandas.DataFrame` with CORIOLIS ID of WMO/CYC profile pairs
+
+        Parameters
+        ----------
+        WMO: int, list(int)
+            Define the list of Argo floats. This is a list of integers with WMO float identifiers.
+            WMO is the World Meteorological Organization.
+        CYC: int, list(int)
+            Define the list of cycle numbers to load ID for each Argo floats listed in ``WMO``.
+
+        Returns
+        -------
+        :class:`pandas.DataFrame`
+    """
+    WMO_list = check_wmo(WMO)
+    if CYC is not None:
+        CYC_list = check_cyc(CYC)
+    URIs = [
+        "https://dataselection.euro-argo.eu/api/trajectory/%i" % wmo for wmo in WMO_list
+    ]
+
+    def prec(data, url):
+        # Transform trajectory json to dataframe
+        # See: https://dataselection.euro-argo.eu/swagger-ui.html#!/cycle-controller/getCyclesByPlatformCodeUsingGET
+        WMO = check_wmo(url.split("/")[-1])[0]
+        rows = []
+        for profile in data:
+            keys = [x for x in profile.keys() if x not in ["coordinate"]]
+            meta_row = dict((key, profile[key]) for key in keys)
+            for row in profile["coordinate"]:
+                meta_row[row] = profile["coordinate"][row]
+            meta_row["WMO"] = WMO
+            rows.append(meta_row)
+        return pd.DataFrame(rows)
+
+    fs = httpstore(cache=True)
+    data = fs.open_mfjson(URIs, preprocess=prec, errors="raise", url_follow=True)
+
+    # Merge results (list of dataframe):
+    key_map = {
+        "id": "ID",
+        "lat": "LATITUDE",
+        "lon": "LONGITUDE",
+        "cvNumber": "CYCLE_NUMBER",
+        "level": "level",
+        "WMO": "PLATFORM_NUMBER",
+    }
+    for i, df in enumerate(data):
+        df = df.reset_index()
+        df = df.rename(columns=key_map)
+        df = df[[value for value in key_map.values() if value in df.columns]]
+        data[i] = df
+    df = pd.concat(data, ignore_index=True)
+    df.sort_values(by=["PLATFORM_NUMBER", "CYCLE_NUMBER"], inplace=True)
+    df = df.reset_index(drop=True)
+    # df = df.set_index(["PLATFORM_NUMBER", "CYCLE_NUMBER"])
+    df = df.astype({"ID": int})
+    if CYC is not None:
+        df = pd.concat([df[df["CYCLE_NUMBER"] == cyc] for cyc in CYC_list]).reset_index(
+            drop=True
+        )
+    return df[
+        ["PLATFORM_NUMBER", "CYCLE_NUMBER", "ID", "LATITUDE", "LONGITUDE", "level"]
+    ]
+
+
+def get_ea_profile_page(WMO, CYC=None):
+    """ Return a list of URL
+
+        Parameters
+        ----------
+        WMO: int, list(int)
+            WMO must be an integer or an iterable with elements that can be casted as integers
+        CYC: int, list(int), default (None)
+            CYC must be an integer or an iterable with elements that can be casted as positive integers
+
+        Returns
+        -------
+        list(str)
+
+        See also
+        --------
+        get_coriolis_profile_id
+    """
+    df = get_coriolis_profile_id(WMO, CYC)
+    url = "https://dataselection.euro-argo.eu/cycle/{}"
+    return [url.format(this_id) for this_id in sorted(df["ID"])]
+
+
+class ArgoNVSReferenceTables:
+    """Argo Reference Tables
+
+    Utility function to retrieve Argo Reference Tables from a NVS server
+    By default, this relies on: https://vocab.nerc.ac.uk/collection
+
+    Examples
+    --------
+    >>> R = ArgoNVSReferenceTables()
+    >>> R.valid_ref
+    >>> R.all_tbl_name()
+    >>> R.tbl(3)
+    >>> R.tbl('R09')
+    >>> R.all_tbl()
+
+    """
+    valid_ref = [
+        "R01",
+        "RR2",
+        "RD2",
+        "RP2",
+        "R03",
+        "R04",
+        "R05",
+        "R06",
+        "R07",
+        "R08",
+        "R09",
+        "R10",
+        "R11",
+        "R12",
+        "R13",
+        "R15",
+        "RMC",
+        "RTV",
+        "R16",
+        "R19",
+        "R20",
+        "R21",
+        "R22",
+        "R23",
+        "R24",
+        "R25",
+        "R26",
+        "R27",
+    ]
+    """List of all available Reference Tables"""
+
+    def __init__(self,
+                 nvs="https://vocab.nerc.ac.uk/collection",
+                 cache: bool = True,
+                 cachedir: str = "",
+                 ):
+        """Argo Reference Tables from NVS"""
+        cachedir = OPTIONS["cachedir"] if cachedir == "" else cachedir
+        self.fs = httpstore(cache=cache, cachedir=cachedir)
+        self.nvs = nvs
+
+    def _valid_ref(self, rtid):
+        if rtid not in self.valid_ref:
+            rtid = "R%0.2d" % rtid
+            if rtid not in self.valid_ref:
+                raise ValueError(
+                    "Invalid Argo Reference Table, should be one in: %s"
+                    % ", ".join(self.valid_ref)
+                )
+        return rtid
+
+    def _jsConcept2df(self, data):
+        """Return all skos:Concept as class:`pandas.DataFrame`"""
+        content = {
+            "altLabel": [],
+            "prefLabel": [],
+            "definition": [],
+            "deprecated": [],
+            "id": [],
+        }
+        for k in data["@graph"]:
+            if k["@type"] == "skos:Collection":
+                Collection_name = k["alternative"]
+            elif k["@type"] == "skos:Concept":
+                content["altLabel"].append(k["altLabel"])
+                content["prefLabel"].append(k["prefLabel"]["@value"])
+                content["definition"].append(k["definition"]["@value"])
+                content["deprecated"].append(k["deprecated"])
+                content["id"].append(k["@id"])
+        df = pd.DataFrame.from_dict(content)
+        df.name = Collection_name
+        return df
+
+    def _jsCollection(self, data):
+        """Return last skos:Collection information as data"""
+        for k in data["@graph"]:
+            if k["@type"] == "skos:Collection":
+                name = k["alternative"]
+                desc = k["description"]
+                rtid = k["@id"]
+        return (name, desc, rtid)
+
+    def get_url(self, rtid, fmt="ld+json"):
+        rtid = self._valid_ref(rtid)
+        if fmt == "ld+json":
+            fmt_ext = "?_profile=nvs&_mediatype=application/ld+json"
+        url = "{}/{}/current/{}".format
+        return url(self.nvs, rtid, fmt_ext)
+
+    def tbl(self, rtid):
+        """Return an Argo Reference table
+
+        Parameters
+        ----------
+        rtid: {str, int}
+            Name or number of the reference table to retrieve. This can be 'R01' or '12'
+
+        Returns
+        -------
+        class:`pandas.DataFrame`
+        """
+        rtid = self._valid_ref(rtid)
+        js = self.fs.open_json(self.get_url(rtid))
+        df = self._jsConcept2df(js)
+        return df
+
+    def tbl_name(self, rtid):
+        """Return name of an Argo Reference table
+
+        Parameters
+        ----------
+        rtid: {str, int}
+            Name or number of the reference table to retrieve. This can be 'R01' or '12'
+
+        Returns
+        -------
+        tuple('short name', 'description', 'NVS id link')
+        """
+        rtid = self._valid_ref(rtid)
+        js = self.fs.open_json(self.get_url(rtid))
+        return self._jsCollection(js)
+
+    def all_tbl(self):
+        """Return all Argo Reference tables
+
+        Returns
+        -------
+        OrderedDict
+            Dictionary with all table short names as key and table content as class:`pandas.DataFrame`
+        """
+        URLs = [self.get_url(rtid) for rtid in self.valid_ref]
+        df_list = self.fs.open_mfjson(URLs, preprocess=self._jsConcept2df)
+        all_tables = {}
+        [all_tables.update({t.name: t}) for t in df_list]
+        all_tables = collections.OrderedDict(sorted(all_tables.items()))
+        return all_tables
+
+    def all_tbl_name(self):
+        """Return names of all Argo Reference tables
+
+        Returns
+        -------
+        OrderedDict
+            Dictionary with all table short names as key and table names as tuple('short name', 'description', 'NVS id link')
+        """
+        URLs = [self.get_url(rtid) for rtid in self.valid_ref]
+        name_list = self.fs.open_mfjson(URLs, preprocess=self._jsCollection)
+        all_tables = {}
+        [
+            all_tables.update({rtid.split("/")[-3]: (name, desc, rtid)})
+            for name, desc, rtid in name_list
+        ]
+        all_tables = collections.OrderedDict(sorted(all_tables.items()))
+        return all_tables