+++ docs/whats-new.rst
@@ -12,7 +12,7 @@ v0.1.XX (X XXX. 2022)
 
 **Features and front-end API**
 
-- New data source ``gdac`` to retrieve data from a GDAC compliant FTP-like source, for DataFetcher and IndexFetcher. You can specify the FTP source with the ``ftp`` fetcher option or with the argopy global option ``ftp``. The FTP source support http, ftp or local files protocols. This fetcher is optimised if pyarrow is available, otherwise pandas dataframe are used. See update on :ref:`Data sources`. (:pr:`157`) by `G. Maze <http://www.github.com/gmaze>`_
+- **New data source ``gdac`` to retrieve data from a GDAC compliant source**, for DataFetcher and IndexFetcher. You can specify the FTP source with the ``ftp`` fetcher option or with the argopy global option ``ftp``. The FTP source support http, ftp or local files protocols. This fetcher is optimised if pyarrow is available, otherwise pandas dataframe are used. See update on :ref:`Data sources`. (:pr:`157`) by `G. Maze <http://www.github.com/gmaze>`_
 
 .. code-block:: python
 
@@ -34,6 +34,44 @@ v0.1.XX (X XXX. 2022)
     Since the new ``gdac`` fetcher can use a local copy of the GDAC ftp server, the legacy ``localftp`` fetcher is now deprecated.
     Using it will raise a warning up to v0.1.12. It will then raise an error in v0.1.13 and will be removed afterward.
 
+- **New dashboard for profiles and new 3rd party dashboards**. Calling on the data fetcher dashboard method will return the Euro-Argo profile page for a single profile. Very useful to look at the data before load. This comes with 2 new utilities functions to get Coriolis ID of profiles (:meth:`utilities.get_coriolis_profile_id`) and to return the list of profile webpages (:meth:`utilities.get_ea_profile_page`). (:pr:`198`) by `G. Maze <http://www.github.com/gmaze>`_.
+
+.. code-block:: python
+
+    from argopy import DataFetcher as ArgoDataFetcher
+    ArgoDataFetcher().profile(5904797, 11).dashboard()
+
+.. code-block:: python
+
+    from argopy.utilities import get_coriolis_profile_id, get_ea_profile_page
+    get_coriolis_profile_id([6902755, 6902756], [11, 12])
+    get_ea_profile_page([6902755, 6902756], [11, 12])
+
+The new profile dashboard can also be accessed with:
+
+.. code-block:: python
+
+    import argopy
+    argopy.dashboard(5904797, 11)
+
+We added the Ocean-OPS (former JCOMMOPS) dashboard for all floats and the Argo-BGC dashboard for BGC floats:
+
+.. code-block:: python
+
+    import argopy
+    argopy.dashboard(5904797, type='ocean-ops')
+    # or
+    argopy.dashboard(5904797, 12, type='bgc')
+
+- New utility function :class:`argopy.utilities.ArgoNVSReferenceTables` to retrieve Argo Reference Tables. (:commit:`cc8fdbe132874b71b35203053626cc29ae7d19c4`) by `G. Maze <http://www.github.com/gmaze>`_.
+
+.. code-block:: python
+
+    from argopy.utilities import ArgoNVSReferenceTables
+    R = ArgoNVSReferenceTables()
+    R.all_tbl_name()
+    R.tbl(3)
+    R.tbl('R09')
 
 **Internals**
 
@@ -63,6 +101,10 @@ v0.1.XX (X XXX. 2022)
 **Breaking changes**
 
 - Index fetcher for local FTP no longer support the option ``index_file``. The name of the file index is internally determined using the dataset requested: ``ar_index_global_prof.txt`` for ``ds='phy'`` and ``argo_synthetic-profile_index.txt`` for ``ds='bgc'``. Using this option will raise a deprecation warning up to v0.1.12 and will then raise an error. (:pr:`157`) by `G. Maze <http://www.github.com/gmaze>`_
+
+- Complete refactoring of the ``argopy.plotters`` module into ``argopy.plot``. (:pr:`198`) by `G. Maze <http://www.github.com/gmaze>`_.
+
+- Remove deprecation warnings for: 'plotters.plot_dac', 'plotters.plot_profilerType'. These now raise an error.
 
 
 v0.1.10 (4 Mar. 2022)